Created useMobileNav custom hook
